fix: guard reconnecting message call
